+++ src/AKS.Application/UseCases/Customers/Transport/CustomerDto.cs
@@ -8,5 +8,5 @@ public record CustomerDto(Guid Id,
     string? State,
     string? Country,
     string? ZipCode,
-    long Latitude,
-    long Longitude);
+    double Latitude,
+    double Longitude);
